[Execution] re-organizing uploader code in Execution engine
